Num/Nat -> i128/u128 (#12)
* Num/Nat -> i128/u128

* fix typo

+++ src/algebra/modint.rs
@@ -8,10 +8,10 @@ use crate::mint; // IGNORE
 use crate::monoid; // IGNORE
 use crate::ring; // IGNORE
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub struct ModInt(pub i64, pub i64); // (residual, modulo)
+pub struct ModInt(pub i128, pub i128); // (residual, modulo)
 
-pub const MOD_1000000007: i64 = 1_000_000_007;
-pub const MOD_998244353: i64 = 998_244_353;
+pub const MOD_1000000007: i128 = 1_000_000_007;
+pub const MOD_998244353: i128 = 998_244_353;
 #[macro_export]
 macro_rules! mint {
     ($x:expr) => {
@@ -20,7 +20,7 @@ macro_rules! mint {
 }
 
 impl ModInt {
-    pub fn new(residual: i64, modulo: i64) -> ModInt {
+    pub fn new(residual: i128, modulo: i128) -> ModInt {
         if residual >= modulo {
             ModInt(residual % modulo, modulo)
         } else if residual < 0 {
@@ -29,11 +29,11 @@ impl ModInt {
             ModInt(residual, modulo)
         }
     }
-    pub fn unwrap(self) -> i64 {
+    pub fn unwrap(self) -> i128 {
         self.0
     }
     pub fn inv(self) -> Self {
-        fn exgcd(r0: i64, a0: i64, b0: i64, r: i64, a: i64, b: i64) -> (i64, i64, i64) {
+        fn exgcd(r0: i128, a0: i128, b0: i128, r: i128, a: i128, b: i128) -> (i128, i128, i128) {
             if r > 0 {
                 exgcd(r, a, b, r0 % r, a0 - r0 / r * a, b0 - r0 / r * b)
             } else {
@@ -46,7 +46,7 @@ impl ModInt {
         }
         ModInt(((a % self.1) + self.1) % self.1, self.1)
     }
-    pub fn pow(self, n: i64) -> Self {
+    pub fn pow(self, n: i128) -> Self {
         if n < 0 {
             self.pow(-n).inv()
         } else if n == 0 {
@@ -90,71 +90,71 @@ ring! {
 }
 impl Field for ModInt {}
 
-impl std::ops::Add<i64> for ModInt {
+impl std::ops::Add<i128> for ModInt {
     type Output = Self;
-    fn add(self, other: i64) -> Self {
+    fn add(self, other: i128) -> Self {
         ModInt::new(self.0 + other, self.1)
     }
 }
-impl std::ops::Add<ModInt> for i64 {
+impl std::ops::Add<ModInt> for i128 {
     type Output = ModInt;
     fn add(self, other: ModInt) -> ModInt {
         other + self
     }
 }
-impl std::ops::AddAssign<i64> for ModInt {
-    fn add_assign(&mut self, other: i64) {
+impl std::ops::AddAssign<i128> for ModInt {
+    fn add_assign(&mut self, other: i128) {
         self.0 = ModInt::new(self.0 + other, self.1).0;
     }
 }
-impl std::ops::Sub<i64> for ModInt {
+impl std::ops::Sub<i128> for ModInt {
     type Output = Self;
-    fn sub(self, other: i64) -> Self {
+    fn sub(self, other: i128) -> Self {
         ModInt::new(self.0 - other, self.1)
     }
 }
-impl std::ops::Sub<ModInt> for i64 {
+impl std::ops::Sub<ModInt> for i128 {
     type Output = ModInt;
     fn sub(self, other: ModInt) -> ModInt {
         ModInt::new(self - other.0, other.1)
     }
 }
-impl std::ops::SubAssign<i64> for ModInt {
-    fn sub_assign(&mut self, other: i64) {
+impl std::ops::SubAssign<i128> for ModInt {
+    fn sub_assign(&mut self, other: i128) {
         self.0 = ModInt::new(self.0 - other, self.1).0;
     }
 }
-impl std::ops::Mul<i64> for ModInt {
+impl std::ops::Mul<i128> for ModInt {
     type Output = Self;
-    fn mul(self, other: i64) -> Self {
+    fn mul(self, other: i128) -> Self {
         ModInt::new(self.0 * other, self.1)
     }
 }
-impl std::ops::Mul<ModInt> for i64 {
+impl std::ops::Mul<ModInt> for i128 {
     type Output = ModInt;
     fn mul(self, other: ModInt) -> ModInt {
         other * self
     }
 }
-impl std::ops::MulAssign<i64> for ModInt {
-    fn mul_assign(&mut self, other: i64) {
+impl std::ops::MulAssign<i128> for ModInt {
+    fn mul_assign(&mut self, other: i128) {
         self.0 = ModInt::new(self.0 * other, self.1).0;
     }
 }
-impl std::ops::Div<i64> for ModInt {
+impl std::ops::Div<i128> for ModInt {
     type Output = Self;
-    fn div(self, other: i64) -> Self {
+    fn div(self, other: i128) -> Self {
         self / ModInt::new(other, self.1)
     }
 }
-impl std::ops::Div<ModInt> for i64 {
+impl std::ops::Div<ModInt> for i128 {
     type Output = ModInt;
     fn div(self, other: ModInt) -> ModInt {
         other.inv() * self
     }
 }
-impl std::ops::DivAssign<i64> for ModInt {
-    fn div_assign(&mut self, other: i64) {
+impl std::ops::DivAssign<i128> for ModInt {
+    fn div_assign(&mut self, other: i128) {
         *self /= ModInt(other, self.1);
     }
 }
